Move nftables_api.go from capture to builder pkg
This PR moves the NftablesAPI apis from the capture pkg
to the builder pkg so that it can be re-used in the ambient
implementation.
